# Conflicts:
#	ime/artifact.legio
#	ime/ime.iml

+++ ime/src/io/intino/ime/box/dsls/LanguageLoader.java
@@ -2,8 +2,8 @@ package io.intino.ime.box.dsls;
 
 import io.intino.alexandria.logger.Logger;
 import io.intino.tara.Language;
-import org.antlr.v4.runtime.misc.Pair;
 import org.apache.commons.codec.binary.Base64;
+import org.apache.commons.lang3.tuple.Pair;
 
 import java.io.File;
 import java.io.IOException;
@@ -15,7 +15,6 @@ import java.nio.file.Files;
 import java.util.HashMap;
 import java.util.Map;
 
-import static io.intino.itrules.formatters.StringFormatters.firstUpperCase;
 
 public class LanguageLoader {
 	public static final String LANGUAGES_PACKAGE = "tara/dsl";
@@ -49,7 +48,7 @@ public class LanguageLoader {
 			String[] parts = dsl.split(":");
 			final ClassLoader classLoader = createClassLoader(jar);
 			if (classLoader == null) return null;
-			Class<?> cls = classLoader.loadClass(LANGUAGES_PACKAGE.replace("/", ".") + "." + firstUpperCase().format(parts[0]).toString());
+			Class<?> cls = classLoader.loadClass(LANGUAGES_PACKAGE.replace("/", ".") + "." + firstUpperCase(parts[0]));
 			return (Language) cls.getConstructors()[0].newInstance();
 		} catch (ClassNotFoundException | InstantiationException | IllegalAccessException |
 				 InvocationTargetException e) {
@@ -60,19 +59,19 @@ public class LanguageLoader {
 
 	private static ClassLoader createClassLoader(File jar) {
 		try {
-			return new URLClassLoader(new URL[]{jar.toURI().toURL()}, io.intino.tara.builder.semantic.LanguageLoader.class.getClassLoader());
+			return new URLClassLoader(new URL[]{jar.toURI().toURL()}, Language.class.getClassLoader());
 		} catch (MalformedURLException e) {
 			Logger.error(e.getMessage(), e);
 			return null;
 		}
 	}
 
-
 	public File download(String dsl) throws IOException {
 		String[] parts = dsl.split(":");
-		String filename = parts[0] + "-" + parts[1] + ".jar";
-		URL url = URI.create(String.join("/", artifactory.toString(), LANGUAGES_PACKAGE, parts[0], parts[1], filename)).toURL();
+		String filename = parts[0] + "-" + parts[1].toLowerCase() + ".jar";
 		File file = new File(dslsDirectory, filename);
+		if (file.exists()) return file;
+		URL url = URI.create(String.join("/", artifactory.toString(), LANGUAGES_PACKAGE, parts[0], parts[1], filename)).toURL();
 		Files.write(file.toPath(), connect(url).readAllBytes());
 		return file;
 	}
@@ -83,7 +82,7 @@ public class LanguageLoader {
 			connection.setConnectTimeout(2000);
 			connection.setReadTimeout(2000);
 			if (credential != null) {
-				String auth = credential.a + ":" + credential.b;
+				String auth = credential.getKey() + ":" + credential.getValue();
 				byte[] encodedAuth = Base64.encodeBase64(auth.getBytes(StandardCharsets.UTF_8));
 				String authHeaderValue = "Basic " + new String(encodedAuth);
 				connection.setRequestProperty("Authorization", authHeaderValue);
@@ -105,5 +104,9 @@ public class LanguageLoader {
 		} catch (Throwable e) {
 			return InputStream.nullInputStream();
 		}
+	}
+
+	private String firstUpperCase(String value) {
+		return value.isEmpty() ? "" : value.substring(0, 1).toUpperCase() + value.substring(1);
 	}
 }
